fix rubocop issues

+++ db/migrate/20251124151349_convert_resource_score_lang_to_language_id.rb
@@ -25,11 +25,13 @@ class ConvertResourceScoreLangToLanguageId < ActiveRecord::Migration[7.1]
       WHERE resource_scores.language_id = languages.id
     SQL
 
-    remove_index :resource_scores,
-      [:resource_id, :language_id, :country] if index_exists?(
-        :resource_scores,
+    if index_exists?(
+      :resource_scores,
+      [:resource_id, :language_id, :country]
+    )
+      remove_index :resource_scores,
         [:resource_id, :language_id, :country]
-      )
+    end
 
     remove_column :resource_scores, :language_id if column_exists?(
       :resource_scores,
